frontend/gcc_dep/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
bcrypt==3.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.3.1
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-fernet-fields==0.6
django-widget-tweaks==1.4.8
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
paramiko==2.7.2
pep517==0.8.2
progress==1.5
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.6
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
